refactor: move more ac setup out of patchwork-isolation
